server/license: move license functions from utilccl to pkg/server/license
Move EnterpriseLicense setting, GetLicense, GetLicenseType, GetLicenseTTL,
GetLicenseEnvironment, and RegisterCallbackOnLicenseChange from
pkg/ccl/utilccl to pkg/server/license. This eliminates the need for the
RegisterCallbackOnLicenseChange init hook.

Informs #164337.
Release note: None
