+++ pkg/ccl/utilccl/BUILD.bazel
@@ -9,47 +9,22 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/ccl/utilccl",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/server/license",
-        "//pkg/server/license/licensepb",
-        "//pkg/settings",
-        "//pkg/settings/cluster",
-        "//pkg/sql/pgwire/pgcode",
-        "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/sem/builtins/builtinsregistry",
         "//pkg/sql/sem/tree",
-        "//pkg/util/log",
-        "//pkg/util/timeutil",
-        "@com_github_cockroachdb_errors//:errors",
     ],
 )
 
 go_test(
     name = "utilccl_test",
     size = "small",
-    srcs = [
-        "license_check_test.go",
-        "license_test.go",
-        "main_test.go",
-    ],
-    embed = [":utilccl"],
+    srcs = ["main_test.go"],
     deps = [
         "//pkg/base",
         "//pkg/ccl",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/server",
-        "//pkg/server/license",
-        "//pkg/server/license/licensepb",
-        "//pkg/settings",
-        "//pkg/settings/cluster",
-        "//pkg/sql",
-        "//pkg/testutils",
         "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
-        "//pkg/util/leaktest",
         "//pkg/util/randutil",
-        "//pkg/util/stop",
-        "//pkg/util/timeutil",
-        "@com_github_stretchr_testify//require",
     ],
 )
