+++ pkg/server/diagnostics/BUILD.bazel
@@ -13,12 +13,12 @@ go_library(
     deps = [
         "//pkg/base",
         "//pkg/build",
-        "//pkg/ccl/utilccl",
         "//pkg/config/zonepb",
         "//pkg/keys",
         "//pkg/kv",
         "//pkg/roachpb",
         "//pkg/server/diagnostics/diagnosticspb",
+        "//pkg/server/license",
         "//pkg/server/license/licensepb",
         "//pkg/server/status",
         "//pkg/server/status/statuspb",
@@ -65,11 +65,11 @@ go_test(
     deps = [
         "//pkg/base",
         "//pkg/build",
-        "//pkg/ccl/utilccl",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/server",
         "//pkg/server/diagnostics/diagnosticspb",
+        "//pkg/server/license",
         "//pkg/server/license/licensepb",
         "//pkg/settings/cluster",
         "//pkg/testutils/diagutils",
