+++ pkg/server/diagnostics/reporter.go
@@ -18,12 +18,12 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/build"
-	"github.com/cockroachdb/cockroach/pkg/ccl/utilccl"
 	"github.com/cockroachdb/cockroach/pkg/config/zonepb"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/kv"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/server/diagnostics/diagnosticspb"
+	"github.com/cockroachdb/cockroach/pkg/server/license"
 	"github.com/cockroachdb/cockroach/pkg/server/license/licensepb"
 	"github.com/cockroachdb/cockroach/pkg/server/status/statuspb"
 	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
@@ -164,7 +164,7 @@ func shouldReportDiagnostics(ctx context.Context, st *cluster.Settings) bool {
 		return true
 	}
 
-	license, err := utilccl.GetLicense(st)
+	license, err := license.GetLicense(st)
 	// If we cannot fetch the license, we do not send the report.
 	if err != nil {
 		log.Dev.Errorf(ctx, "error fetching license in shouldReportDiagnostics: %s", err)
@@ -228,7 +228,7 @@ func (r *Reporter) ReportDiagnostics(ctx context.Context) {
 
 	report := r.CreateReport(ctx, telemetry.ResetCounts)
 
-	license, err := utilccl.GetLicense(r.Settings)
+	license, err := license.GetLicense(r.Settings)
 	if err != nil {
 		if log.V(2) {
 			log.Dev.Warningf(ctx, "failed to retrieve license while reporting diagnostics: %v", err)
