+++ pkg/server/license/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
     srcs = [
         "cclbridge.go",
         "enforcer.go",
+        "license.go",
         "opts.go",
         ":gen-lictype-stringer",  # keep
     ],
@@ -13,6 +14,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/keys",
+        "//pkg/server/license/licensepb",
+        "//pkg/settings",
         "//pkg/settings/cluster",
         "//pkg/sql/isql",
         "//pkg/sql/pgwire/pgcode",
@@ -32,19 +35,26 @@ go_test(
     name = "license_test",
     srcs = [
         "enforcer_test.go",
+        "license_check_test.go",
+        "license_decode_test.go",
+        "license_enforcer_integration_test.go",
         "main_test.go",
     ],
+    embed = [":license"],
     deps = [
-        ":license",
         "//pkg/base",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/server",
+        "//pkg/server/license/licensepb",
+        "//pkg/settings",
+        "//pkg/settings/cluster",
         "//pkg/sql",
         "//pkg/sql/catalog/descs",
         "//pkg/sql/isql",
         "//pkg/sql/sessiondata",
         "//pkg/storage/fs",
+        "//pkg/testutils",
         "//pkg/testutils/listenerutil",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/sqlutils",
